+++ app/controllers/publications_controller.rb
@@ -44,12 +44,13 @@ class PublicationsController < ApplicationController
   end
 
   def send_reminder
-    if current_user.secretary? and @publication = Publication.current.find_by_id(params[:id]) and @reviewer = User.current.find_by_id(params[:reviewer_id])
+    @publication = Publication.current.find_by_id(params[:id])
+    @reviewer = User.current.find_by_id(params[:reviewer_id])
+    if @publication && @reviewer && current_user.secretary?
       @user_publication_review = @reviewer.user_publication_reviews.find_by_publication_id(@publication.id)
       @user_publication_review = @reviewer.user_publication_reviews.create(publication_id: @publication.id) if @user_publication_review.blank?
       @user_publication_review.update_column :reminder_sent_at, Time.zone.now
-
-      UserMailer.publication_approval_reminder(current_user, params[:to], params[:cc], params[:subject], params[:body]).deliver_later if Rails.env.production?
+      UserMailer.publication_approval_reminder(current_user, params[:to], params[:cc], params[:subject], params[:body].to_s).deliver_later if EMAILS_ENABLED
     else
       render nothing: true
     end
@@ -84,7 +85,7 @@ class PublicationsController < ApplicationController
 
       extension = params[:publication][:manuscript].original_filename.downcase.split('.').last
       message = ManuscriptUploader.new.extension_white_list.include?(extension) ? nil : "Not a valid document type: #{extension}"
-      flash[:notice] = "Manuscript was successfully uploaded." if message.blank?
+      flash[:notice] = 'Manuscript was successfully uploaded.' if message.blank?
 
       redirect_to @publication, alert: message
       # render 'publications/manuscripts/show_manuscript'
@@ -109,9 +110,10 @@ class PublicationsController < ApplicationController
   end
 
   def pp_approval
-    if current_user.pp_committee_secretary? and @publication = Publication.current.find_by_id(params[:id]) and params[:publication]
+    @publication = Publication.current.find_by_id(params[:id])
+    if @publication && current_user.pp_committee_secretary? && params[:publication]
       @publication.update(params.require(:publication).permit(:status, :manuscript_number, :additional_ppcommittee_instructions))
-      UserMailer.publication_approval(@publication, true, current_user).deliver_later if @publication.status != 'proposed' and @publication.user and Rails.env.production?
+      UserMailer.publication_approval(@publication, true, current_user).deliver_later if @publication.status != 'proposed' && @publication.user && EMAILS_ENABLED
       @publication.send_reminders(current_user) if @publication.status == 'approved'
       redirect_to @publication
     else
@@ -128,9 +130,10 @@ class PublicationsController < ApplicationController
   end
 
   def sc_approval
-    if current_user.steering_committee_secretary? and @publication = Publication.current.find_by_id(params[:id]) and params[:publication]
+    @publication = Publication.current.find_by_id(params[:id])
+    if @publication && current_user.steering_committee_secretary? && params[:publication]
       @publication.update(params.require(:publication).permit(:status, :additional_sccommittee_instructions))
-      UserMailer.publication_approval(@publication, false, current_user).deliver_later if @publication.status != 'approved' and @publication.user and Rails.env.production?
+      UserMailer.publication_approval(@publication, false, current_user).deliver_later if @publication.status != 'approved' && @publication.user && EMAILS_ENABLED
       redirect_to @publication
     else
       redirect_to root_path