+++ app/controllers/users_controller.rb
@@ -4,8 +4,7 @@
 class UsersController < ApplicationController
   before_action :authenticate_user!
   before_action :check_system_admin, only: [:new, :create, :edit, :update, :destroy]
-  before_action :set_user, only: [ :show, :edit, :update, :destroy ]
-  before_action :redirect_without_user, only: [ :show, :edit, :update, :destroy ]
+  before_action :set_user, only: [:show, :edit, :update, :destroy]
 
   def index
     unless current_user.system_admin? or params[:format] == 'json'
@@ -41,10 +40,10 @@ class UsersController < ApplicationController
   # Post /users/activate.json
   def activate
     params[:user] ||= {}
-    params[:user][:password] = params[:user][:password_confirmation] = Digest::SHA1.hexdigest(Time.zone.now.usec.to_s)[0..19] if params[:user][:password].blank? and params[:user][:password_confirmation].blank?
+    params[:user][:password] = params[:user][:password_confirmation] = Digest::SHA1.hexdigest(Time.zone.now.usec.to_s)[0..19] if params[:user][:password].blank? && params[:user][:password_confirmation].blank?
     @user = User.new(user_params)
     if @user.save
-      UserMailer.status_activated(@user).deliver_later if Rails.env.production? and @user.status == 'active'
+      UserMailer.status_activated(@user).deliver_later if EMAILS_ENABLED && @user.status == 'active'
 
       respond_to do |format|
         format.html { redirect_to @user, notice: 'User was successfully created.' }
@@ -61,7 +60,7 @@ class UsersController < ApplicationController
   def update
     original_status = @user.status
     if @user.update(user_params)
-      UserMailer.status_activated(@user).deliver_later if Rails.env.production? and original_status != @user.status and @user.status == 'active'
+      UserMailer.status_activated(@user).deliver_later if EMAILS_ENABLED && original_status != @user.status && @user.status == 'active'
       redirect_to @user, notice: 'User was successfully updated.'
     else
       render action: 'edit'
@@ -75,20 +74,20 @@ class UsersController < ApplicationController
 
   private
 
-    def set_user
-      @user = User.current.find_by_id(params[:id])
-    end
+  def set_user
+    @user = User.current.find_by_id(params[:id])
+    redirect_without_user
+  end
 
-    def redirect_without_user
-      empty_response_or_root_path(users_path) unless @user
-    end
+  def redirect_without_user
+    empty_response_or_root_path(users_path) unless @user
+  end
 
-    def user_params
-      params[:user] ||= {}
-
-      params.require(:user).permit(
-        :first_name, :last_name, :email, :password, :password_confirmation, :pp_committee, :pp_committee_secretary, :steering_committee, :steering_committee_secretary, :system_admin, :status
-      )
-    end
-
+  def user_params
+    params.require(:user).permit(
+      :first_name, :last_name, :email, :password, :password_confirmation,
+      :pp_committee, :pp_committee_secretary, :steering_committee,
+      :steering_committee_secretary, :system_admin, :status
+    )
+  end
 end