+++ app/mailers/application_mailer.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+# Generic mailer setup
+class ApplicationMailer < ActionMailer::Base
+  default from: "#{ENV['website_name']} <#{ActionMailer::Base.smtp_settings[:email]}>"
+  layout 'mailer'
+
+  protected
+
+  def setup_email
+  #   attachments.inline['logo.png'] = File.read('app/assets/images/chat.png')
+  # rescue
+  #   nil
+  end
+end