+++ config/initializers/action_mailer.rb
@@ -1,16 +1,21 @@
-ActionMailer::Base.default_url_options[:host] = ENV['website_url'].to_s.gsub(/^https?:\/\//, '')
-unless Rails.env.test?
+# frozen_string_literal: true
+
+ActionMailer::Base.default_url_options[:host] = ENV['website_url'].to_s.gsub(%r{^https?://}, '')
+if Rails.env.test?
+  ActionMailer::Base.delivery_method = :test
+  ActionMailer::Base.smtp_settings[:email] = 'travis-ci@example.com'
+  EMAILS_ENABLED = true
+else
   ActionMailer::Base.delivery_method = :smtp
   ActionMailer::Base.smtp_settings = {
     enable_starttls_auto: true,
     address: ENV['smtp_address'],
     port: ENV['smtp_port'].to_i,
-    domain: "",
+    domain: '',
     authentication: ENV['smtp_authentication'].to_sym, # :plain, :login, or, :cram_md5
     email: ENV['smtp_email'],
     user_name: ENV['smtp_user_name'],
     password: ENV['smtp_password']
   }
-else
-  ActionMailer::Base.smtp_settings[:email] = "travis-ci@example.com"
+  EMAILS_ENABLED = (ENV['emails_enabled'] == 'true')
 end