+++ config/routes.rb
@@ -1,5 +1,6 @@
-Rails.application.routes.draw do
+# frozen_string_literal: true
 
+Rails.application.routes.draw do
   resources :publications do
     collection do
       get :search
@@ -7,25 +8,37 @@ Rails.application.routes.draw do
       get :print_latex
     end
     member do
-       post :pp_approval
-       post :sc_approval
-       post :show_subcommittee_decision
-       post :edit_subcommittee_decision
-       post :show_steering_committee_decision
-       post :edit_steering_committee_decision
-       post :upload_manuscript
-       post :edit_manuscript
-       post :show_manuscript
-       post :destroy_manuscript
-       post :send_reminder
-       post :inline_update
-       post :tag_for_review
-       post :remove_nomination
-       post :archive
+      post :pp_approval
+      post :sc_approval
+      post :show_subcommittee_decision
+      post :edit_subcommittee_decision
+      post :show_steering_committee_decision
+      post :edit_steering_committee_decision
+      post :upload_manuscript
+      post :edit_manuscript
+      post :show_manuscript
+      post :destroy_manuscript
+      post :send_reminder
+      post :inline_update
+      post :tag_for_review
+      post :remove_nomination
+      post :archive
     end
   end
 
-  devise_for :users, controllers: { registrations: 'contour/registrations', sessions: 'contour/sessions', passwords: 'contour/passwords', confirmations: 'contour/confirmations', unlocks: 'contour/unlocks' }, path_names: { sign_up: 'register', sign_in: 'login' }, path: ""
+  devise_for :users,
+             controllers: {
+               registrations: 'contour/registrations',
+               sessions: 'contour/sessions',
+               passwords: 'contour/passwords',
+               confirmations: 'contour/confirmations',
+               unlocks: 'contour/unlocks'
+             },
+             path_names: {
+               sign_up: 'register',
+               sign_in: 'login'
+             },
+             path: ''
 
   resources :users do
     collection do
@@ -40,8 +53,5 @@ Rails.application.routes.draw do
     get :version
   end
 
-  root to: "publications#index"
-
-  # See how all your routes lay out with "rake routes"
-
+  root to: 'publications#index'
 end