+++ test/controllers/publications_controller_test.rb
@@ -2,6 +2,7 @@
 
 require 'test_helper'
 
+# Tests to assure publication proposals can be created, updated, and viewed.
 class PublicationsControllerTest < ActionController::TestCase
   setup do
     @proposed = publications(:proposed)
@@ -9,7 +10,7 @@ class PublicationsControllerTest < ActionController::TestCase
     @nominated = publications(:nominated)
   end
 
-  test "should tag for sc review" do
+  test 'should tag for sc review' do
     login(users(:sc_secretary))
     post :tag_for_review, id: @proposed, committee: 'sc', tagged: '1', format: 'js'
     assert_not_nil assigns(:publication)
@@ -18,7 +19,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'tag_for_review'
   end
 
-  test "should untag for sc review" do
+  test 'should untag for sc review' do
     login(users(:sc_secretary))
     post :tag_for_review, id: @proposed, committee: 'sc', format: 'js'
     assert_not_nil assigns(:publication)
@@ -27,7 +28,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'tag_for_review'
   end
 
-  test "should tag for pp review" do
+  test 'should tag for pp review' do
     login(users(:pp_secretary))
     post :tag_for_review, id: @proposed, committee: 'pp', tagged: '1', format: 'js'
     assert_not_nil assigns(:publication)
@@ -36,7 +37,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'tag_for_review'
   end
 
-  test "should untag for pp review" do
+  test 'should untag for pp review' do
     login(users(:pp_secretary))
     post :tag_for_review, id: @proposed, committee: 'pp', format: 'js'
     assert_not_nil assigns(:publication)
@@ -45,58 +46,59 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'tag_for_review'
   end
 
-  test "should not tag for review as non-secretary" do
+  test 'should not tag for review as non-secretary' do
     login(users(:valid))
     post :tag_for_review, id: @proposed, committee: 'pp', tagged: '1', format: 'js'
     assert_nil assigns(:publication)
     assert_response :success
   end
 
-  test "should send reminder" do
+  test 'should send reminder' do
     login(users(:pp_secretary))
-    post :send_reminder, id: @proposed, reviewer_id: users(:pp_committee).to_param, format: 'js'
+    post :send_reminder, id: @proposed, reviewer_id: users(:pp_committee).to_param, to: 'to@example.com', cc: 'cc@example.com', format: 'js'
     assert_not_nil assigns(:publication)
     assert_not_nil assigns(:reviewer)
     assert_not_nil assigns(:user_publication_review)
     assert_template 'send_reminder'
+    assert_response :success
   end
 
-  test "should not send reminder for non-secretary" do
+  test 'should not send reminder for non-secretary' do
     login(users(:valid))
     post :send_reminder, id: @proposed, reviewer_id: users(:pp_committee).to_param, format: 'js'
-    assert_nil assigns(:publication)
-    assert_nil assigns(:reviewer)
+    assert_not_nil assigns(:publication)
+    assert_not_nil assigns(:reviewer)
     assert_nil assigns(:user_publication_review)
     assert_response :success
   end
 
-  test "should edit manuscript" do
+  test 'should edit manuscript' do
     login(users(:valid))
     post :edit_manuscript, id: @nominated, format: 'js'
     assert_not_nil assigns(:publication)
     assert_template 'publications/manuscripts/edit_manuscript'
   end
 
-  test "should not edit manuscript on publication pending sc approval" do
+  test 'should not edit manuscript on publication pending sc approval' do
     login(users(:valid))
     post :edit_manuscript, id: @proposed, format: 'js'
     assert_response :success
   end
 
-  test "should show manuscript" do
+  test 'should show manuscript' do
     login(users(:valid))
     post :show_manuscript, id: @nominated, format: 'js'
     assert_not_nil assigns(:publication)
     assert_template 'publications/manuscripts/show_manuscript'
   end
 
-  test "should not show manuscript on publication pending sc approval" do
+  test 'should not show manuscript on publication pending sc approval' do
     login(users(:valid))
     post :show_manuscript, id: @proposed, format: 'js'
     assert_response :success
   end
 
-  test "should upload manuscript in DOC format" do
+  test 'should upload manuscript in DOC format' do
     login(users(:valid))
     post :upload_manuscript, id: @nominated, publication: { manuscript: fixture_file_upload('../../test/support/publications/manuscripts/test_01.doc') }
     assert_not_nil assigns(:publication)
@@ -104,7 +106,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should upload manuscript in DOCX format" do
+  test 'should upload manuscript in DOCX format' do
     login(users(:valid))
     post :upload_manuscript, id: @nominated, publication: { manuscript: fixture_file_upload('../../test/support/publications/manuscripts/test_01.docx') }
     assert_not_nil assigns(:publication)
@@ -112,7 +114,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should upload manuscript in PDF format" do
+  test 'should upload manuscript in PDF format' do
     login(users(:valid))
     post :upload_manuscript, id: @nominated, publication: { manuscript: fixture_file_upload('../../test/support/publications/manuscripts/test_01.pdf') }
     assert_not_nil assigns(:publication)
@@ -120,7 +122,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should not upload manuscript on publication pending sc approval" do
+  test 'should not upload manuscript on publication pending sc approval' do
     login(users(:valid))
     post :upload_manuscript, id: @proposed, publication: { manuscript: fixture_file_upload('../../test/support/publications/manuscripts/test_01.doc') }
     assert_not_nil assigns(:publication)
@@ -128,7 +130,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should not upload manuscript non-existent manuscript" do
+  test 'should not upload manuscript non-existent manuscript' do
     login(users(:valid))
     post :upload_manuscript, id: @nominated
     assert_not_nil assigns(:publication)
@@ -136,119 +138,119 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should not upload manuscript" do
+  test 'should not upload manuscript' do
     login(users(:two))
     post :upload_manuscript, id: @nominated, publication: { manuscript: fixture_file_upload('../../test/support/publications/manuscripts/test_01.doc') }
     assert_nil assigns(:publication)
     assert_redirected_to root_path
   end
 
-  test "should destroy manuscript" do
+  test 'should destroy manuscript' do
     login(users(:valid))
     post :destroy_manuscript, id: @proposed, format: 'js'
     assert_not_nil assigns(:publication)
     assert_template 'publications/manuscripts/edit_manuscript'
   end
 
-  test "should not destroy manuscript" do
+  test 'should not destroy manuscript' do
     login(users(:two))
     post :destroy_manuscript, id: @proposed, format: 'js'
     assert_nil assigns(:publication)
     assert_response :success
   end
 
-  test "should send pp approval" do
+  test 'should send pp approval' do
     login(users(:pp_secretary))
     post :pp_approval, id: @proposed, publication: { status: 'approved', additional_sccommittee_instructions: 'Additional Instructions' }
     assert_not_nil assigns(:publication)
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should not send pp approval for non-pp committee secretary" do
+  test 'should not send pp approval for non-pp committee secretary' do
     login(users(:valid))
     post :pp_approval, id: @proposed, publication: { status: 'approved', additional_sccommittee_instructions: 'Additional Instructions' }
-    assert_nil assigns(:publication)
+    assert_not_nil assigns(:publication)
     assert_redirected_to root_path
   end
 
-  test "should send sc approval" do
+  test 'should send sc approval' do
     login(users(:sc_secretary))
     post :sc_approval, id: @proposed, publication: { status: 'nominated', additional_sccommittee_instructions: 'Additional Instructions' }
     assert_not_nil assigns(:publication)
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should not send sc approval for non-steering committee secretary" do
+  test 'should not send sc approval for non-steering committee secretary' do
     login(users(:valid))
     post :sc_approval, id: @proposed, publication: { status: 'nominated', additional_sccommittee_instructions: 'Additional Instructions' }
-    assert_nil assigns(:publication)
+    assert_not_nil assigns(:publication)
     assert_redirected_to root_path
   end
 
-  test "should show subcommittee decision" do
+  test 'should show subcommittee decision' do
     login(users(:pp_secretary))
     post :show_subcommittee_decision, id: @proposed, format: 'js'
     assert_not_nil assigns(:publication)
     assert_template 'show_subcommittee_decision'
   end
 
-  test "should not show subcommittee decision for non sc secretary" do
+  test 'should not show subcommittee decision for non sc secretary' do
     login(users(:valid))
     post :show_subcommittee_decision, id: @proposed, format: 'js'
     assert_nil assigns(:publication)
     assert_response :success
   end
 
-  test "should get edit subcommittee decision" do
+  test 'should get edit subcommittee decision' do
     login(users(:pp_secretary))
     post :edit_subcommittee_decision, id: @proposed, format: 'js'
     assert_not_nil assigns(:publication)
     assert_template 'edit_subcommittee_decision'
   end
 
-  test "should not get edit subcommittee decision for non sc secretary" do
+  test 'should not get edit subcommittee decision for non sc secretary' do
     login(users(:valid))
     post :edit_subcommittee_decision, id: @proposed, format: 'js'
     assert_nil assigns(:publication)
     assert_response :success
   end
 
-  test "should show steering committee decision" do
+  test 'should show steering committee decision' do
     login(users(:sc_secretary))
     post :show_steering_committee_decision, id: @proposed, format: 'js'
     assert_not_nil assigns(:publication)
     assert_template 'show_steering_committee_decision'
   end
 
-  test "should not show steering committee decision for non sc secretary" do
+  test 'should not show steering committee decision for non sc secretary' do
     login(users(:valid))
     post :show_steering_committee_decision, id: @proposed, format: 'js'
     assert_nil assigns(:publication)
     assert_response :success
   end
 
-  test "should get edit steering committee decision" do
+  test 'should get edit steering committee decision' do
     login(users(:sc_secretary))
     post :edit_steering_committee_decision, id: @proposed, format: 'js'
     assert_not_nil assigns(:publication)
     assert_template 'edit_steering_committee_decision'
   end
 
-  test "should not get edit steering committee decision for non sc secretary" do
+  test 'should not get edit steering committee decision for non sc secretary' do
     login(users(:valid))
     post :edit_steering_committee_decision, id: @proposed, format: 'js'
     assert_nil assigns(:publication)
     assert_response :success
   end
 
-  test "should get index" do
+  test 'should get index' do
     login(users(:valid))
     get :index
     assert_not_nil assigns(:publications)
     assert_response :success
   end
 
-  test "should get print" do
+  test 'should get print' do
     login(users(:valid))
     get :print
     assert_not_nil assigns(:order)
@@ -256,7 +258,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not show 'draft' or 'not approved' for regular users on the publication matrix unless they are the author" do
+  test 'should not show `draft` or `not approved` for regular users on the publication matrix unless they are the author' do
     login(users(:two))
     get :index, format: 'js'
     assert_not_nil assigns(:publications)
@@ -264,7 +266,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'index'
   end
 
-  test "should sort marked for P&P review to top in index if user P&P Secretary" do
+  test 'should sort marked for P&P review to top in index if user P&P Secretary' do
     login(users(:pp_secretary))
     get :index
     assert_response :success
@@ -273,7 +275,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_equal tagged_array, tagged_array.sort.reverse
   end
 
-  test "should sort marked for P&P review to top in index if user P&P Member" do
+  test 'should sort marked for P&P review to top in index if user P&P Member' do
     login(users(:pp_committee))
     get :index
     assert_response :success
@@ -282,7 +284,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_equal tagged_array, tagged_array.sort.reverse
   end
 
-  test "should sort marked for SC review to top in index if user SC Secretary" do
+  test 'should sort marked for SC review to top in index if user SC Secretary' do
     login(users(:sc_secretary))
     get :index
     assert_response :success
@@ -291,7 +293,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_equal tagged_array, tagged_array.sort.reverse
   end
 
-  test "should sort marked for SC review to top in index if user SC Member" do
+  test 'should sort marked for SC review to top in index if user SC Member' do
     login(users(:sc_committee))
     get :index
     assert_response :success
@@ -300,7 +302,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_equal tagged_array, tagged_array.sort.reverse
   end
 
-  test "should sort marked for P&P review followed by SC review to top in index if user is in both P&P and SC committees" do
+  test 'should sort marked for P&P review followed by SC review to top in index if user is in both P&P and SC committees' do
     login(users(:pp_and_sc_committee))
     get :index
     assert_response :success
@@ -309,7 +311,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_equal tagged_array, tagged_array.sort.reverse
   end
 
-  test "should sort marked for P&P review followed by SC review to top in index if user is in both P&P and SC secretary" do
+  test 'should sort marked for P&P review followed by SC review to top in index if user is in both P&P and SC secretary' do
     login(users(:pp_and_sc_secretary))
     get :index
     assert_response :success
@@ -318,13 +320,13 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_equal tagged_array, tagged_array.sort.reverse
   end
 
-  test "should get new" do
+  test 'should get new' do
     login(users(:valid))
     get :new
     assert_response :success
   end
 
-  test "should update inline for secretary" do
+  test 'should update inline for secretary' do
     login(users(:pp_secretary))
     post :inline_update, id: @proposed, publication: { targeted_start_date: "05/20/2011", status: 'submitted' }, format: 'js'
     assert_not_nil assigns(:publication)
@@ -334,14 +336,14 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not update inline for non-secretary" do
+  test 'should not update inline for non-secretary' do
     login(users(:valid))
     post :inline_update, id: @proposed, publication: { targeted_start_date: "05/20/2011", status: 'submitted' }, format: 'js'
     assert_nil assigns(:publication)
     assert_response :success
   end
 
-  test "should create publication" do
+  test 'should create publication' do
     login(users(:valid))
     assert_difference('Publication.count') do
       post :create, publication: @draft.attributes
@@ -350,7 +352,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should create publication using quick save" do
+  test 'should create publication using quick save' do
     login(users(:valid))
     assert_difference('Publication.count') do
       post :create, publication: @draft.attributes, publish: -1
@@ -360,7 +362,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'new'
   end
 
-  test "should not create publication draft without full title and abbreviated title" do
+  test 'should not create publication draft without full title and abbreviated title' do
     login(users(:valid))
     assert_difference('Publication.count', 0) do
       post :create, publication: @draft.attributes.merge({ full_title: '', abbreviated_title: ''})
@@ -372,7 +374,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'new'
   end
 
-  test "should create publication for secretary" do
+  test 'should create publication for secretary' do
     login(users(:pp_secretary))
     assert_difference('Publication.count') do
       post :create, :publication => @draft.attributes
@@ -381,44 +383,44 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should show publication" do
+  test 'should show publication' do
     login(users(:valid))
     get :show, id: @proposed
     assert_response :success
   end
 
-  test "should show publication without a lead author" do
+  test 'should show publication without a lead author' do
     login(users(:pp_secretary))
     get :show, id: publications(:no_lead_author)
     assert_response :success
   end
 
-  test "should show publication with committee reminders" do
+  test 'should show publication with committee reminders' do
     login(users(:sc_secretary))
     get :show, id: publications(:three)
     assert_response :success
   end
 
-  test "should get edit if publication is a draft" do
+  test 'should get edit if publication is a draft' do
     login(users(:valid))
     get :edit, id: @draft
     assert_response :success
   end
 
-  test "should not get edit if publication is proposed" do
+  test 'should not get edit if publication is proposed' do
     login(users(:valid))
     get :edit, id: @proposed
     assert_redirected_to root_path
   end
 
-  test "should get edit if publication is proposed and secretary wishes to edit" do
+  test 'should get edit if publication is proposed and secretary wishes to edit' do
     login(users(:pp_secretary))
     get :edit, id: @proposed
     assert_template 'edit'
     assert_response :success
   end
 
-  test "should update publication and submit for review" do
+  test 'should update publication and submit for review' do
     login(users(:valid))
     put :update, id: @draft, publication: @proposed.attributes, publish: 1
     assert_not_nil assigns(:publication)
@@ -426,7 +428,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should update publication and save as a draft" do
+  test 'should update publication and save as a draft' do
     login(users(:valid))
     put :update, id: @draft, publication: @proposed.attributes
     assert_not_nil assigns(:publication)
@@ -434,7 +436,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should not update publication draft without full title and abbreviated title" do
+  test 'should not update publication draft without full title and abbreviated title' do
     login(users(:valid))
     put :update, id: @draft, publication: @proposed.attributes.merge({ full_title: '', abbreviated_title: ''})
     assert_not_nil assigns(:publication)
@@ -444,14 +446,14 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'edit'
   end
 
-  test "should not update publication with invalid id" do
+  test 'should not update publication with invalid id' do
     login(users(:valid))
     put :update, id: -1, publication: @proposed.attributes
     assert_nil assigns(:publication)
     assert_redirected_to publications_path
   end
 
-  test "should update publication and quick save" do
+  test 'should update publication and quick save' do
     login(users(:valid))
     put :update, id: @draft, publication: @proposed.attributes, publish: -1
     assert_not_nil assigns(:publication)
@@ -459,7 +461,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'edit'
   end
 
-  test "should update publication as secretary" do
+  test 'should update publication as secretary' do
     login(users(:pp_secretary))
     put :update, id: @proposed, publication: { status: 'nominated' }
     assert_not_nil assigns(:publication)
@@ -467,7 +469,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publication_path(assigns(:publication))
   end
 
-  test "should destroy publication" do
+  test 'should destroy publication' do
     login(users(:valid))
     assert_difference('Publication.current.count', -1) do
       delete :destroy, id: @draft
@@ -476,7 +478,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publications_path
   end
 
-  test "should not destroy publication with invalid id" do
+  test 'should not destroy publication with invalid id' do
     login(users(:valid))
     assert_difference('Publication.current.count', 0) do
       delete :destroy, id: -1
@@ -486,7 +488,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_redirected_to publications_path
   end
 
-  test "should remove nomination from publication reviews" do
+  test 'should remove nomination from publication reviews' do
     login(users(:pp_secretary))
     post :remove_nomination, id: publications(:tagged_for_pp_and_sc_review), nomination: 'Joe Schmoe', format: 'js'
 
@@ -495,7 +497,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_template 'committee_nominations'
   end
 
-  test "should not remove nomination from publication reviews for non-secretary" do
+  test 'should not remove nomination from publication reviews for non-secretary' do
     login(users(:valid))
     post :remove_nomination, id: publications(:tagged_for_pp_and_sc_review), nomination: 'Joe Schmoe', format: 'js'
 
@@ -504,7 +506,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not remove nomination from publication reviews for invalid publication" do
+  test 'should not remove nomination from publication reviews for invalid publication' do
     login(users(:pp_secretary))
     post :remove_nomination, id: -1, nomination: 'Joe Schmoe', format: 'js'
 
@@ -512,7 +514,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should archive publication" do
+  test 'should archive publication' do
     login(users(:pp_secretary))
     post :archive, id: @proposed, undo: 'false', format: 'js'
     assert_not_nil assigns(:publication)
@@ -521,7 +523,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should unarchive publication" do
+  test 'should unarchive publication' do
     login(users(:pp_secretary))
     post :archive, id: @proposed, undo: 'true', format: 'js'
     assert_not_nil assigns(:publication)
@@ -530,7 +532,7 @@ class PublicationsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not archive publication for non-secretary" do
+  test 'should not archive publication for non-secretary' do
     login(users(:valid))
     post :archive, id: @proposed, undo: 'false', format: 'js'
     assert_nil assigns(:publication)